Add configuration Xresources for urxvt

+++ setup.sh
@@ -12,6 +12,14 @@ DOTFILES="$(dirname $(readlink -f $0))"
 
 # Ubuntu 18.04 locations
 VIMHOME=${HOME}/.vim
+XRESOURCES=${HOME}/.Xresources
 
-ln -s ${DOTFILES}/vim ${VIMHOME}
-ln -s ${DOTFILES}/tmux/tmux.conf ${HOME}/.tmux.conf
+if [ ! -e ${VIMHOME} ]; then
+    ln -s ${DOTFILES}/vim ${VIMHOME}
+fi
+if [ ! -e ${HOME}/.tmux.conf ]; then
+    ln -s ${DOTFILES}/tmux/tmux.conf ${HOME}/.tmux.conf
+fi
+if [ ! -e ${XRESOURCES} ]; then
+    ln -s ${DOTFILES}/Xresources/gruvbox-dark.xresources ${XRESOURCES}
+fi
